DataWriter Update
Added a bunch of methods for the DataWriter, it can now read specific
lines as well as reading all integers found on a line, even though there
may also be non-numeric characters on the line. As well as make the
DataFile extend File which opens up a lot of opportunity for the class.